sun4i_csi1: fix up ctrl setting
V4L2_CTRL_TYPE_U32 vs V4L2_CTRL_TYPE_INTEGER was an interesting one...

U32 makes it a compound type, for some reason, meaning that it is
treated differently, and that it should be handled through the union
v4l2_ctrl_ptr at the end of the v4l2_ctrl struct. I think this type was
added with EXT_CTRL functionality or something, and can therefore only
be used with those calls. Very non-intuitive.

Signed-off-by: Luc Verhaegen <libv@skynet.be>